-fixed small warning message when compiling (I just dont like warning messages lol) -changed coverflow color detection again, it's hopefully a bit more efficient now

+++ source/gui/text.cpp
@@ -483,7 +483,7 @@ void Asciify( wchar_t *str )
 
 void Asciify2( char *str )
 {
-	int i=0;
+	u8 i=0;
 	for( i=0; i < strlen(str); ++i )
 	{
 		if( str[i] < 0x20 || str[i] > 0x7F )
